\name{Study_FDR}
\alias{Study_FDR}
\title{Generate the BH part of Figure 2 of paper}
\usage{
Study_FDR(nsim)
}
\arguments{
  \item{nsim}{Number of simulations}
}
\value{Array with values of average power}
\description{
Computes the average power of BH and conditional BH an array of input values as defined in Figure 2 of the paper.
}
\examples{
Study_FDR(100)
}
