\name{Study_Power_Fig3}
\alias{Study_Power_Fig3}
\title{Study the power in a range of situations, as displayed in figure 3}
\usage{
Study_Power_Fig3(nsim)
}
\arguments{
  \item{nsim}{Number of simulations}
}
\value{Array with values of minimum power}
\description{
Computes the minimum power of various multiple testing procedures, as defined in Figure 3 of the paper.
}
\examples{
Study_Power_Fig3(100)
}
